refactor: split tui/update por domínio
