+++ src/tui/update/mod.rs
@@ -1,829 +1,87 @@
-use ratatui::crossterm::event::{KeyCode, KeyEvent};
+//! Pure update: mutates AppState from Action, returns follow-up actions.
+//!
+//! Split por domínio (navigation / config / login / fetch / history).
+//! `pub fn update` é o único entry point público.
 
-use super::action::Action;
-use super::mouse::{ChipKind, MouseTarget};
-use super::state::{
-    sidebar_items, AppState, ConfigField, ConfigState, FetchStatus, FxEvent, HistoryRange,
-    ProviderView, Screen, SidebarItem,
-};
+mod config;
+mod fetch;
+mod history;
+mod login;
+mod navigation;
 
-/// Translates a KeyEvent into a semantic Action using current screen state:
-/// up/down move the sidebar cursor, Enter/h/g/w activate a sidebar item
-/// (jumping directly to Detail/History/Login/Waybar). Esc goes back to
-/// Detail from History/Login/Waybar (Task 11: Detail é a tela default —
-/// Esc nela é no-op, ver o match genérico abaixo). Some screens (Waybar
-/// editing, Login list) intercept keys with their own local semantics
-/// before falling through to this navigation.
-fn key_to_action_with_state(key: KeyEvent, state: &AppState) -> Option<Action> {
-    // Se o overlay de ajuda esta aberto, qualquer tecla fecha (Esc ou '?').
-    if state.show_help {
-        return match key.code {
-            KeyCode::Esc | KeyCode::Char('?') | KeyCode::Char('q') => Some(Action::ToggleHelp),
-            _ => None,
-        };
-    }
-
-    // '?' global: abre o overlay de ajuda de qualquer contexto (exceto edicao).
-    let in_config_edit = state.screen == Screen::Waybar
-        && state
-            .config_state
-            .as_ref()
-            .map(|cs| cs.editing)
-            .unwrap_or(false);
-
-    if key.code == KeyCode::Char('?') && !in_config_edit {
-        return Some(Action::ToggleHelp);
-    }
-
-    // Na tela Waybar com campo em edicao, delega ao input buffer (so Esc/Enter escapam).
-    if state.screen == Screen::Waybar {
-        if let Some(cs) = &state.config_state {
-            if cs.editing {
-                return match key.code {
-                    KeyCode::Esc => Some(Action::ConfigCancelEdit),
-                    KeyCode::Enter => Some(Action::ConfigConfirmEdit),
-                    _ => None, // o event_loop vai passar o evento cru ao Input
-                };
-            }
-        }
-        // Tela Waybar, fora do modo edicao. h/g/w saltam direto para outra
-        // tela (substituem o antigo ←→ de troca de aba — sem conflito com
-        // j/k/Enter/s/Esc/q, ja usados aqui).
-        return match key.code {
-            KeyCode::Char('j') | KeyCode::Down => Some(Action::ConfigDown),
-            KeyCode::Char('k') | KeyCode::Up => Some(Action::ConfigUp),
-            KeyCode::Enter => Some(Action::ConfigEnterEdit),
-            KeyCode::Char('s') => Some(Action::SaveConfig),
-            KeyCode::Esc => Some(Action::Back),
-            KeyCode::Char('h') => Some(Action::Activate(SidebarItem::History)),
-            KeyCode::Char('g') => Some(Action::Activate(SidebarItem::Login)),
-            KeyCode::Char('w') => Some(Action::Activate(SidebarItem::Waybar)),
-            KeyCode::Char('q') => Some(Action::Quit),
-            _ => None,
-        };
-    }
-
-    // Tela Login: navegacao e acao de login. h/w saltam direto para outra
-    // tela (mesmo racional da tela Waybar acima).
-    if state.screen == Screen::Login {
-        return match key.code {
-            KeyCode::Char('j') | KeyCode::Down => Some(Action::LoginDown),
-            KeyCode::Char('k') | KeyCode::Up => Some(Action::LoginUp),
-            KeyCode::Enter => Some(Action::LoginRequested(
-                login_selected_id(state.login_selected).to_string(),
-            )),
-            KeyCode::Esc => Some(Action::Back),
-            KeyCode::Char('h') => Some(Action::Activate(SidebarItem::History)),
-            KeyCode::Char('g') => Some(Action::Activate(SidebarItem::Login)),
-            KeyCode::Char('w') => Some(Action::Activate(SidebarItem::Waybar)),
-            KeyCode::Char('q') => Some(Action::Quit),
-            _ => None,
-        };
-    }
-
-    // Tela History: navega a lista de dias (j/k/Enter) + alterna o range do
-    // chart (t) — mesmo racional da tela Login acima (h/g/w saltam direto
-    // pra outra tela; j/k/Enter aqui SUBSTITUEM a navegação genérica de
-    // sidebar/Detail, exatamente como LoginUp/LoginDown/LoginRequested
-    // substituem Down/Up/OpenDetail na tela Login).
-    if state.screen == Screen::History {
-        return match key.code {
-            KeyCode::Char('j') | KeyCode::Down => Some(Action::HistoryDown),
-            KeyCode::Char('k') | KeyCode::Up => Some(Action::HistoryUp),
-            KeyCode::Enter => Some(Action::HistoryToggleDay),
-            KeyCode::Char('t') => Some(Action::ToggleHistoryRange),
-            KeyCode::Char('r') => Some(Action::Refresh),
-            KeyCode::Esc => Some(Action::Back),
-            KeyCode::Char('h') => Some(Action::Activate(SidebarItem::History)),
-            KeyCode::Char('g') => Some(Action::Activate(SidebarItem::Login)),
-            KeyCode::Char('w') => Some(Action::Activate(SidebarItem::Waybar)),
-            KeyCode::Char('q') => Some(Action::Quit),
-            _ => None,
-        };
-    }
-
-    // Tela Detail (default, sem bloco dedicado acima): Esc é no-op de
-    // propósito (Task 11) — Detail é a base da navegação, não há "voltar"
-    // daqui (o overlay de ajuda já intercepta Esc antes de chegar aqui).
-    match key.code {
-        KeyCode::Char('j') | KeyCode::Down => Some(Action::Down),
-        KeyCode::Char('k') | KeyCode::Up => Some(Action::Up),
-        KeyCode::Enter => Some(Action::OpenDetail),
-        KeyCode::Char('h') => Some(Action::Activate(SidebarItem::History)),
-        KeyCode::Char('g') => Some(Action::Activate(SidebarItem::Login)),
-        KeyCode::Char('w') => Some(Action::Activate(SidebarItem::Waybar)),
-        KeyCode::Char('r') => Some(Action::Refresh),
-        KeyCode::Char('q') => Some(Action::Quit),
-        _ => None,
-    }
-}
-
-/// Mapeia o índice selecionado da aba Login pro id do provider — mesma
-/// ordem de `render/login.rs::PROVIDERS`. Compartilhado entre o Enter
-/// (`key_to_action_with_state`) e o clique no chip `StartLogin` (T14):
-/// os dois precisam disparar a MESMA action pro provider selecionado.
-fn login_selected_id(idx: usize) -> &'static str {
-    match idx {
-        0 => "claude",
-        1 => "codex",
-        _ => "amp",
-    }
-}
-
-/// Retorna a representacao em string de um campo das edit_settings.
-fn field_value_string(field: ConfigField, cs: &ConfigState) -> String {
-    let s = &cs.edit_settings;
-    match field {
-        ConfigField::Providers => s.waybar.providers.join(", "),
-        ConfigField::ProviderOrder => s.waybar.provider_order.join(", "),
-        ConfigField::Separators => format!("{:?}", s.waybar.separators).to_lowercase(),
-        ConfigField::DisplayMode => format!("{:?}", s.waybar.display_mode).to_lowercase(),
-        ConfigField::FxRate => format!("{:.2}", s.fx_rate),
-        ConfigField::Signal => s
-            .waybar
-            .signal
-            .map(|n| n.to_string())
-            .unwrap_or_else(|| "none".to_string()),
-        ConfigField::Interval => s.waybar.interval.to_string(),
-    }
-}
-
-/// Aplica o valor textual do buffer de edicao ao campo das edit_settings.
-/// Retorna Err com mensagem descritiva se o valor e invalido.
-fn apply_field_edit(field: ConfigField, value: &str, cs: &mut ConfigState) -> Result<(), String> {
-    let s = &mut cs.edit_settings;
-    match field {
-        ConfigField::Providers => {
-            let providers: Vec<String> = value
-                .split(',')
-                .map(|p| p.trim().to_string())
-                .filter(|p| !p.is_empty())
-                .collect();
-            let (normalized, order) =
-                crate::settings::normalize_provider_selection(&providers, &s.waybar.provider_order);
-            s.waybar.providers = normalized;
-            s.waybar.provider_order = order;
-            Ok(())
-        }
-        ConfigField::ProviderOrder => {
-            let order: Vec<String> = value
-                .split(',')
-                .map(|p| p.trim().to_string())
-                .filter(|p| !p.is_empty())
-                .collect();
-            let (normalized, new_order) =
-                crate::settings::normalize_provider_selection(&s.waybar.providers, &order);
-            s.waybar.providers = normalized;
-            s.waybar.provider_order = new_order;
-            Ok(())
-        }
-        ConfigField::Separators => {
-            use crate::settings::SeparatorStyle;
-            let sep = match value.trim() {
-                "pill" => SeparatorStyle::Pill,
-                "gap" => SeparatorStyle::Gap,
-                "bare" => SeparatorStyle::Bare,
-                "glass" => SeparatorStyle::Glass,
-                "shadow" => SeparatorStyle::Shadow,
-                "none" => SeparatorStyle::None,
-                other => {
-                    return Err(format!(
-                        "separador inválido: '{other}' (pill/gap/bare/glass/shadow/none)"
-                    ))
-                }
-            };
-            s.waybar.separators = sep;
-            Ok(())
-        }
-        ConfigField::DisplayMode => {
-            use crate::settings::DisplayMode;
-            let mode = match value.trim() {
-                "remaining" => DisplayMode::Remaining,
-                "used" => DisplayMode::Used,
-                other => return Err(format!("modo inválido: '{other}' (remaining/used)")),
-            };
-            s.waybar.display_mode = mode;
-            Ok(())
-        }
-        ConfigField::FxRate => {
-            let rate: f64 = value
-                .trim()
-                .parse()
-                .map_err(|_| format!("fxRate inválido: '{value}' (número esperado)"))?;
-            if !rate.is_finite() || rate <= 0.0 {
-                return Err(format!("fxRate deve ser positivo: '{value}'"));
-            }
-            s.fx_rate = rate;
-            Ok(())
-        }
-        ConfigField::Signal => {
-            let trimmed = value.trim();
-            if trimmed == "none" || trimmed.is_empty() {
-                s.waybar.signal = None;
-                return Ok(());
-            }
-            let n: i64 = trimmed
-                .parse()
-                .map_err(|_| format!("signal inválido: '{value}' (1-30 ou none)"))?;
-            if !(1..=30).contains(&n) {
-                return Err(format!("signal deve ser 1-30 ou none: '{value}'"));
-            }
-            s.waybar.signal = Some(n as u8);
-            Ok(())
-        }
-        ConfigField::Interval => {
-            let n: u32 = value
-                .trim()
-                .parse()
-                .map_err(|_| format!("interval inválido: '{value}' (inteiro positivo)"))?;
-            if n == 0 {
-                return Err(format!("interval deve ser > 0: '{value}'"));
-            }
-            s.waybar.interval = n;
-            Ok(())
-        }
-    }
-}
+use crate::tui::action::Action;
+use crate::tui::state::AppState;
 
 /// Pure update function: mutates `state` based on `action`, returns follow-up actions.
 /// No IO, no spawning, no clocks — fully testable.
 pub fn update(state: &mut AppState, action: Action) -> Vec<Action> {
     match action {
         Action::Key(key) => {
-            if let Some(semantic) = key_to_action_with_state(key, state) {
+            if let Some(semantic) = navigation::key_to_action_with_state(key, state) {
                 return update(state, semantic);
             }
             vec![]
         }
 
-        Action::Up => {
-            state.sidebar_selected = state.sidebar_selected.saturating_sub(1);
-            vec![]
-        }
+        Action::Up => navigation::up(state),
+        Action::Down => navigation::down(state),
+        Action::OpenDetail => navigation::open_detail(state),
+        Action::Activate(item) => navigation::activate(state, item),
+        Action::SelectSidebar(i) => navigation::select_sidebar(state, i),
+        Action::Back => navigation::back(state),
+        Action::Quit => navigation::quit(state),
+        Action::ToggleHelp => navigation::toggle_help(state),
+        Action::Tick => navigation::tick(),
+        Action::AnimTick => navigation::anim_tick(state),
+        Action::Click(target) => navigation::click(state, target),
+        Action::Hover(t) => navigation::hover(state, t),
+        Action::Scroll(delta) => navigation::scroll(state, delta),
 
-        Action::Down => {
-            let max = sidebar_items(state.providers.len()).len() - 1;
-            state.sidebar_selected = (state.sidebar_selected + 1).min(max);
-            vec![]
-        }
-
-        Action::OpenDetail => {
-            // Recursa (mesmo padrao do braço Action::Key): aplica a ativacao
-            // na hora e propaga os follow-ups dela (ex. InitConfig ao ativar
-            // Waybar), em vez de so devolver Activate sem aplicar.
-            let items = sidebar_items(state.providers.len());
-            match items.get(state.sidebar_selected).copied() {
-                Some(item) => update(state, Action::Activate(item)),
-                None => vec![],
-            }
-        }
-
-        Action::Activate(item) => {
-            let old_screen = state.screen;
-            let follow_ups = match item {
-                SidebarItem::Provider(i) => {
-                    state.selected = i;
-                    state.screen = Screen::Detail;
-                    vec![]
-                }
-                SidebarItem::History => {
-                    state.screen = Screen::History;
-                    vec![]
-                }
-                SidebarItem::Login => {
-                    state.screen = Screen::Login;
-                    vec![]
-                }
-                SidebarItem::Waybar => {
-                    state.screen = Screen::Waybar;
-                    // Inicializa config_state ao entrar na tela Waybar pela 1a vez.
-                    // Nao podemos acessar settings aqui (update e puro), entao sinalizamos
-                    // com um placeholder — o event_loop (drain) sobrescreve com as
-                    // settings reais ao interceptar InitConfig (mesmo mecanismo do
-                    // antigo SwitchTab para a aba Waybar).
-                    if state.config_state.is_none() {
-                        vec![Action::InitConfig(crate::settings::Settings {
-                            version: 0, // sentinela; event_loop sobrescreve com real
-                            waybar: crate::settings::Waybar {
-                                providers: vec![],
-                                show_percentage: true,
-                                separators: crate::settings::SeparatorStyle::Gap,
-                                provider_order: vec![],
-                                display_mode: crate::settings::DisplayMode::Remaining,
-                                signal: None,
-                                interval: 60,
-                            },
-                            tooltip: crate::settings::Tooltip {},
-                            models: Default::default(),
-                            window_policy: Default::default(),
-                            notify: crate::settings::Notify { enabled: true },
-                            cache: crate::settings::CacheSettings {
-                                ttl: Default::default(),
-                            },
-                            menu: crate::settings::MenuSettings {
-                                animations: true,
-                                font_family: "IBM Plex Mono".to_string(),
-                                font_size: 12,
-                            },
-                            glyph_mode: crate::settings::GlyphMode::Box,
-                            fx_rate: 5.50,
-                        })]
-                    } else {
-                        vec![]
-                    }
-                }
-            };
-            // Sincroniza o cursor da sidebar com o item ativado — cobre
-            // h/g/w, Enter (via OpenDetail) e cliques futuros da Task 9.
-            // Sem isso, ativar por atalho deixa sidebar_selected apontando
-            // pra outro item quando a Task 10 desenhar o highlight.
-            if let Some(idx) = sidebar_items(state.providers.len())
-                .iter()
-                .position(|i| *i == item)
-            {
-                state.sidebar_selected = idx;
-            }
-            // `state.scroll` é compartilhado entre telas (ScrollView do
-            // Overview e, agora, a tabela do History) — sem reset, uma
-            // posição de scroll deixada numa tela (ex. Overview rolado)
-            // vaza pra outra tela sem relação nenhuma com aquele offset
-            // (History abriria com linhas de dado já puladas). Reseta
-            // só quando a tela de fato muda, não quando reativa a que
-            // já está ativa (ex. Enter em cima de si mesmo).
-            if state.screen != old_screen {
-                state.scroll = 0;
-            }
-            follow_ups
-        }
-
-        Action::SelectSidebar(i) => {
-            state.sidebar_selected = i;
-            vec![]
-        }
-
-        Action::Back => {
-            // Mesma regra do braço Activate: mudança de tela zera o scroll
-            // (compartilhado entre Detail/History) — só quando a screen de
-            // fato muda, senão Esc numa tela que já é Detail zeraria o
-            // scroll do usuário à toa. Task 11: Detail (não mais Overview) é
-            // o destino — Overview morreu, Detail é a tela default/home.
-            if state.screen != Screen::Detail {
-                state.scroll = 0;
-            }
-            state.screen = Screen::Detail;
-            vec![]
-        }
-
-        Action::Refresh => {
-            // Evita fetch duplicado se ja tem um em voo; senao, re-enfileira
-            // Refresh UMA vez para o event_loop interceptar (mesmo padrao de
-            // ReloadUsage/SaveConfig — o drain NAO re-entra no update com ele).
-            if state.fetch_pending.is_empty() {
-                vec![Action::Refresh]
-            } else {
-                vec![]
-            }
-        }
-
-        Action::FetchStarted(ids) => {
-            state.status = FetchStatus::Loading;
-            // Uniao sem duplicatas (nao overwrite): ondas de fetch podem se
-            // sobrepor (Refresh/tick de 60s disparados enquanto uma onda
-            // anterior ainda resolve). Se um id ja esta pendente, a propria
-            // onda em voo vai resolve-lo de novo — nao precisa re-adicionar.
-            for id in ids {
-                if !state.fetch_pending.contains(&id) {
-                    state.fetch_pending.push(id);
-                }
-            }
-            vec![]
-        }
-
-        Action::ProviderFetched(q) => {
-            state.fetch_pending.retain(|id| id != &q.provider);
-            let old_len = state.providers.len();
-            let provider_id = q.provider.clone();
-            match state
-                .providers
-                .iter_mut()
-                .find(|pv| pv.quota.provider == q.provider)
-            {
-                Some(pv) => pv.quota = *q,
-                None => {
-                    state.providers.push(ProviderView::new(*q));
-                    // Provider novo insere Provider(old_len) na posicao
-                    // old_len da sidebar (Task 11: sem Overview, os
-                    // providers começam direto no índice 0 — sem o
-                    // deslocamento de +1 que a Overview antes exigia). Um
-                    // cursor que já apontava pra History/Login/Waybar
-                    // (índices >= old_len) passaria a apontar pro item
-                    // anterior sem isso — desloca 1 posição pra manter o
-                    // cursor no MESMO item lógico.
-                    if state.sidebar_selected >= old_len {
-                        state.sidebar_selected += 1;
-                    }
-                }
-            }
-            // Boot/foco pendente (Task 11): resolve por ID, lazy, nunca por
-            // índice fixo — o fetch de OUTRO provider não pode roubar o
-            // foco (só entra aqui quando o provider ALVO chega). Provider(i)
-            // é o i-ésimo item da sidebar (sem offset de Overview).
-            if state.pending_focus.as_deref() == Some(provider_id.as_str()) {
-                if let Some(idx) = state
-                    .providers
-                    .iter()
-                    .position(|p| p.quota.provider == provider_id)
-                {
-                    state.selected = idx;
-                    state.screen = Screen::Detail;
-                    state.sidebar_selected = idx;
-                    state.pending_focus = None;
-                }
-            }
-            vec![]
-        }
-
+        Action::Refresh => fetch::refresh(state),
+        Action::FetchStarted(ids) => fetch::fetch_started(state, ids),
+        Action::ProviderFetched(q) => fetch::provider_fetched(state, q),
         Action::FetchCompleted { fetched_at, silent } => {
-            // Efeito sweep (T16, fix pós-review): dispara quando uma onda
-            // de fetch termina, mesmo se outra onda sobreposta ainda
-            // estiver em voo (gatilho é a action chegando, não o status
-            // final agregado) — MAS só para ondas pedidas pelo usuário
-            // (load inicial, `r`/chip Refresh, LoginFinished). O poll
-            // silencioso de 60s (`silent=true`) NUNCA dispara o sweep
-            // (spec §8: efeito é feedback de ação, não deve repetir a cada
-            // minuto sem o usuário fazer nada). O resto do bookkeeping
-            // abaixo (last_update monotônico, status, ReloadUsage) é
-            // agnóstico à origem da onda — count-up do header pode
-            // continuar re-lerpando num poll silencioso, é sutil.
-            if !silent {
-                state.fx_queue.push(FxEvent::FetchLanded);
-            }
-            // NAO limpa fetch_pending incondicionalmente: cada ProviderFetched
-            // ja remove o proprio id. Se sobrar algo aqui, e porque outra onda
-            // (Refresh/tick de 60s) ainda esta em voo — mantem Loading em vez
-            // de regredir pra Loaded (a onda que sobrar completa depois).
-            //
-            // Mesmo parse de timestamp usado pelo antigo DataFetched, mas
-            // nunca regride: fica o mais recente entre o atual e o parseado
-            // (ondas sobrepostas podem terminar fora de ordem).
-            let parsed = time::OffsetDateTime::parse(
-                &fetched_at,
-                &time::format_description::well_known::Rfc3339,
-            )
-            .ok();
-            if let Some(new) = parsed {
-                state.last_update = Some(state.last_update.map_or(new, |cur| cur.max(new)));
-            }
-            if state.fetch_pending.is_empty() {
-                state.status = FetchStatus::Loaded;
-            }
-            // Clamp selection if providers list shrank.
-            if !state.providers.is_empty() && state.selected >= state.providers.len() {
-                state.selected = state.providers.len() - 1;
-            }
-            vec![Action::ReloadUsage]
+            fetch::fetch_completed(state, fetched_at, silent)
         }
+        Action::ReloadUsage => fetch::reload_usage(),
+        Action::FetchFailed(msg) => fetch::fetch_failed(state, msg),
+        Action::UsageComputed(summary) => fetch::usage_computed(state, summary),
 
-        Action::ReloadUsage => vec![], // interceptada no event_loop; no update e no-op
+        Action::InitConfig(settings) => config::init_config(state, settings),
+        Action::ConfigUp => config::config_up(state),
+        Action::ConfigDown => config::config_down(state),
+        Action::ConfigEnterEdit => config::config_enter_edit(state),
+        Action::ConfigCancelEdit => config::config_cancel_edit(state),
+        Action::ConfigConfirmEdit => config::config_confirm_edit(state),
+        Action::SaveConfig => config::save_config(state),
+        Action::ConfigSaveResult(result) => config::config_save_result(state, result),
 
-        Action::FetchFailed(msg) => {
-            // Estreitado pela Task 5: so cobre erro de RUNTIME da thread de
-            // `spawn_fetch` (ex. falha ao construir o tokio Builder) — erros
-            // de provider (rede/parse/auth) viajam embutidos no
-            // ProviderQuota.error e chegam via ProviderFetched, nunca aqui.
-            // Limpa fetch_pending: senao a thread morta deixa o spinner
-            // girando pra sempre (nenhum ProviderFetched vai chegar).
-            state.fetch_pending.clear();
-            state.status = FetchStatus::Failed(msg);
-            vec![]
-        }
+        Action::LoginUp => login::login_up(state),
+        Action::LoginDown => login::login_down(state),
+        Action::LoginRequested(id) => login::login_requested(state, id),
+        Action::LoginResult(result) => login::login_result(state, result),
+        Action::LoginFinished(id) => login::login_finished(id),
 
-        Action::Quit => {
-            state.should_quit = true;
-            vec![]
-        }
-
-        Action::UsageComputed(summary) => {
-            // 1º load (usage ainda None): pinta display_cost = alvo direto,
-            // sem animar a partir de zero. Loads seguintes deixam o
-            // AnimTick fazer o count-up até o novo alvo.
-            if state.usage.is_none() {
-                state.display_cost = summary.total_cost.usd;
-            }
-            state.usage = Some(summary);
-            vec![]
-        }
-
-        // --- Config tab actions ---
-        Action::InitConfig(settings) => {
-            // So inicializa se ainda nao existe (preserva edicoes em andamento).
-            if state.config_state.is_none() {
-                state.config_state = Some(ConfigState::new(&settings));
-            } else if let Some(cs) = state.config_state.as_mut() {
-                // Atualiza com as settings reais se ainda era o placeholder (version=0).
-                if cs.edit_settings.version == 0 {
-                    *cs = ConfigState::new(&settings);
-                }
-            }
-            vec![]
-        }
-
-        Action::ConfigUp => {
-            if let Some(cs) = state.config_state.as_mut() {
-                if !cs.editing && cs.selected_field > 0 {
-                    cs.selected_field -= 1;
-                }
-            }
-            vec![]
-        }
-
-        Action::ConfigDown => {
-            if let Some(cs) = state.config_state.as_mut() {
-                let max = ConfigField::ALL.len().saturating_sub(1);
-                if !cs.editing && cs.selected_field < max {
-                    cs.selected_field += 1;
-                }
-            }
-            vec![]
-        }
-
-        Action::ConfigEnterEdit => {
-            if let Some(cs) = state.config_state.as_mut() {
-                if !cs.editing {
-                    let field = ConfigField::ALL[cs.selected_field];
-                    let current = field_value_string(field, cs);
-                    cs.input = tui_input::Input::new(current);
-                    cs.editing = true;
-                    cs.status_msg = None;
-                }
-            }
-            vec![]
-        }
-
-        Action::ConfigCancelEdit => {
-            if let Some(cs) = state.config_state.as_mut() {
-                cs.editing = false;
-                cs.input = tui_input::Input::default();
-            }
-            vec![]
-        }
-
-        Action::ConfigConfirmEdit => {
-            if let Some(cs) = state.config_state.as_mut() {
-                if cs.editing {
-                    let field = ConfigField::ALL[cs.selected_field];
-                    let value = cs.input.value().to_string();
-                    match apply_field_edit(field, &value, cs) {
-                        Ok(()) => {
-                            cs.editing = false;
-                            cs.input = tui_input::Input::default();
-                            cs.status_msg =
-                                Some("Campo atualizado. Pressione [s] para salvar.".to_string());
-                        }
-                        Err(e) => {
-                            cs.status_msg = Some(format!("Erro: {e}"));
-                            // Mantem edicao aberta para correcao.
-                        }
-                    }
-                }
-            }
-            vec![]
-        }
-
-        Action::SaveConfig => {
-            // Sinaliza pending_save: o event_loop pinta "Salvando..." no
-            // frame atual e SO ENTAO faz o IO (persist + reload Waybar) no
-            // topo do proximo loop. O update e puro, nao faz IO.
-            if state.config_state.is_some() {
-                if let Some(cs) = state.config_state.as_mut() {
-                    cs.status_msg = Some("Salvando...".to_string());
-                }
-                state.pending_save = true;
-            }
-            vec![]
-        }
-
-        Action::ConfigSaveResult(result) => {
-            if let Some(cs) = state.config_state.as_mut() {
-                cs.status_msg = Some(match result {
-                    Ok(()) => "Configuracao salva e Waybar recarregado.".to_string(),
-                    Err(e) => format!("Erro ao salvar: {e}"),
-                });
-            }
-            vec![]
-        }
-
-        // --- Aba Login ---
-        Action::LoginUp => {
-            if state.login_selected > 0 {
-                state.login_selected -= 1;
-            }
-            state.login_status = None;
-            vec![]
-        }
-
-        Action::LoginDown => {
-            // 3 providers: indices 0, 1, 2.
-            if state.login_selected < 2 {
-                state.login_selected += 1;
-            }
-            state.login_status = None;
-            vec![]
-        }
-
-        Action::LoginRequested(id) => {
-            // Puro: sinaliza pending_login. O event_loop pinta o status no
-            // frame atual e SO ENTAO suspende o terminal para o CLI de login
-            // (fix: "Abrindo login..." nunca era pintado antes desta task).
-            state.login_status = Some(format!("Abrindo login para {}...", id));
-            state.pending_login = Some(id);
-            vec![]
-        }
-
-        Action::LoginResult(result) => {
-            state.login_status = Some(match result {
-                // Refetch agora e automatico (LoginFinished), sem precisar de [r].
-                Ok(()) => "Login concluido. Atualizando quota...".to_string(),
-                Err(e) => format!("Erro no login: {e}"),
-            });
-            vec![]
-        }
-
-        Action::LoginFinished(id) => {
-            // Re-enfileira UMA vez para o event_loop interceptar (mesmo
-            // padrao de Refresh/ReloadUsage): o drain chama spawn_fetch(only)
-            // direto, sem re-entrar no update com esta action.
-            vec![Action::LoginFinished(id)]
-        }
-
-        Action::HistoryLoaded(records) => {
-            state.history = Some(records);
-            vec![]
-        }
-
-        Action::ToggleHistoryRange => {
-            state.history_range = match state.history_range {
-                HistoryRange::Day => HistoryRange::Week,
-                HistoryRange::Week => HistoryRange::Day,
-            };
-            vec![]
-        }
-
-        Action::HistoryDown => {
-            // `sessions_by_day` é barato o bastante pra recomputar por tecla
-            // (7 dias de records — YAGNI cachear no state por ora). Clamp
-            // aqui, não no render: `history_selected` nunca aponta pra fora
-            // da lista real.
-            let n_days = state
-                .history
-                .as_deref()
-                .map(|r| crate::usage::buckets::sessions_by_day(r, state.local_offset).len())
-                .unwrap_or(0);
-            if n_days > 0 {
-                state.history_selected = (state.history_selected + 1).min(n_days - 1);
-            }
-            vec![]
-        }
-
-        Action::HistoryUp => {
-            state.history_selected = state.history_selected.saturating_sub(1);
-            vec![]
-        }
-
-        Action::HistoryToggleDay => {
-            if let Some(records) = state.history.as_deref() {
-                let days = crate::usage::buckets::sessions_by_day(records, state.local_offset);
-                if let Some(day) = days.get(state.history_selected) {
-                    if !state.history_expanded.remove(&day.date) {
-                        state.history_expanded.insert(day.date);
-                    }
-                }
-            }
-            vec![]
-        }
-
-        Action::ToggleHelp => {
-            state.show_help = !state.show_help;
-            vec![]
-        }
-
-        Action::Tick => vec![],
-
-        Action::AnimTick => {
-            // (A antiga "Animação A" — gauge lerp via display_ratio — foi
-            // removida: nenhum render lia o valor.)
-            // Animação C (throbber): avança o frame do spinner braille.
-            state.throbber.advance();
-            // Animação D (pulse): contador de frames para blink do ● crítico
-            // da sidebar (Task 10). O pulse dos gauges críticos do
-            // card/detalhe morreu em v8 (spec §6, `pulse_color` removido) —
-            // gauge agora é sólido, sem modulação de brilho.
-            state.anim_frame = state.anim_frame.wrapping_add(1);
-            // Count-up do custo do header (T16): ease exponencial (~800ms
-            // em ticks de 30ms, fator 0.12). animations=false → snapa
-            // direto pro alvo, sem lerp visual.
-            let target_cost = state
-                .usage
-                .as_ref()
-                .map(|u| u.total_cost.usd)
-                .unwrap_or(0.0);
-            if state.animations {
-                state.display_cost += (target_cost - state.display_cost) * 0.12;
-                if (target_cost - state.display_cost).abs() < 0.01 {
-                    state.display_cost = target_cost;
-                }
-            } else {
-                state.display_cost = target_cost;
-            }
-            vec![]
-        }
-
-        // --- Mouse (Task 9) ---
-        // Recursa via update() em vez de so devolver a action mapeada como
-        // follow-up (mesmo padrao de Action::Key/Action::OpenDetail acima):
-        // aplica o efeito NA HORA. Alem de manter clique e teclado simetricos
-        // (ex. o Refresh clicado passa pelo MESMO guard anti-fetch-duplicado
-        // que o Refresh da tecla [r] — devolver a action crua ignoraria o
-        // guard, ja que drain() intercepta Action::Refresh direto sem
-        // reentrar no update), a sincronicidade e o que os testes observam.
-        Action::Click(target) => match target {
-            MouseTarget::Sidebar(i) => {
-                // sidebar_items() e a unica fonte de verdade dos indices validos
-                // (SelectSidebar nao tem bounds-check); ignora i fora de faixa em
-                // vez de deixar sidebar_selected apontando pra um item inexistente.
-                let items = sidebar_items(state.providers.len());
-                match items.get(i).copied() {
-                    Some(item) => {
-                        state.sidebar_selected = i;
-                        update(state, Action::Activate(item))
-                    }
-                    None => vec![],
-                }
-            }
-            // Na tela Login, os itens da lista de providers TAMBÉM registram
-            // MouseTarget::Card(i) (mesma linguagem visual dos cards do
-            // Overview) — mas aqui o clique so SELECIONA (mesmo efeito de
-            // LoginUp/LoginDown), nunca ativa Detail. Reusar o braço
-            // genérico abaixo (Activate(Provider(i))) navegaria pra fora da
-            // tela Login no primeiro clique, contrariando "click seleciona;
-            // ativação continua pelo Enter/chip" (T14).
-            MouseTarget::Card(i) if state.screen == Screen::Login => {
-                if i < 3 {
-                    state.login_selected = i;
-                    state.login_status = None;
-                }
-                vec![]
-            }
-            // Cards eram do Overview/dashboard (Task 11: ambos apagados) —
-            // fora da tela Login (guard acima), Card não tem mais nenhum
-            // efeito (a lista de providers da Detail é a sidebar, não
-            // cards clicáveis).
-            MouseTarget::Card(_) => vec![],
-            MouseTarget::Chip(ChipKind::Open) => update(state, Action::OpenDetail),
-            MouseTarget::Chip(ChipKind::Refresh) => update(state, Action::Refresh),
-            MouseTarget::Chip(ChipKind::Help) => update(state, Action::ToggleHelp),
-            MouseTarget::Chip(ChipKind::Quit) => update(state, Action::Quit),
-            MouseTarget::Chip(ChipKind::Back) => update(state, Action::Back),
-            MouseTarget::Chip(ChipKind::Login) => {
-                update(state, Action::Activate(SidebarItem::Login))
-            }
-            MouseTarget::Chip(ChipKind::History) => {
-                update(state, Action::Activate(SidebarItem::History))
-            }
-            MouseTarget::Chip(ChipKind::ToggleRange) => update(state, Action::ToggleHistoryRange),
-            MouseTarget::Chip(ChipKind::ExpandDay) => update(state, Action::HistoryToggleDay),
-            // Chip "iniciar login" da tela Login: dispara a MESMA action que
-            // o Enter dispara lá (Action::LoginRequested pro provider
-            // selecionado) — nunca Activate(Login), que seria no-op (a
-            // tela já está ativa) e ignoraria o clique (T14).
-            MouseTarget::Chip(ChipKind::StartLogin) => update(
-                state,
-                Action::LoginRequested(login_selected_id(state.login_selected).to_string()),
-            ),
-            MouseTarget::Chip(ChipKind::EnterEdit) => update(state, Action::ConfigEnterEdit),
-            MouseTarget::Chip(ChipKind::SaveConfig) => update(state, Action::SaveConfig),
-        },
-
-        Action::Hover(t) => {
-            state.hover = t;
-            vec![]
-        }
-
-        Action::Scroll(delta) => {
-            // saturating_add_signed ja satura em 0 (limite inferior de u16) —
-            // .max(0) e redundante (clippy::unnecessary_min_or_max).
-            state.scroll = state.scroll.saturating_add_signed(delta as i16);
-            vec![]
-        }
+        Action::HistoryLoaded(records) => history::history_loaded(state, records),
+        Action::ToggleHistoryRange => history::toggle_history_range(state),
+        Action::HistoryDown => history::history_down(state),
+        Action::HistoryUp => history::history_up(state),
+        Action::HistoryToggleDay => history::history_toggle_day(state),
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use ratatui::crossterm::event::KeyModifiers;
+    use ratatui::crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 
-    use super::*;
+    use super::update;
     use crate::providers::types::ProviderQuota;
+    use crate::tui::action::Action;
+    use crate::tui::mouse::{ChipKind, MouseTarget};
+    use crate::tui::state::{
+        sidebar_items, AppState, ConfigField, FetchStatus, HistoryRange, ProviderView, Screen,
+        SidebarItem,
+    };
 
     fn fake_quota(id: &str) -> ProviderQuota {
         ProviderQuota {
